Massive final changes
56. increased the maximum reach of each foot during the searching phase.
110/118/126/158/169/206/224. Reincorporated rotation (DO NOT turn
quickly; be careful with it)
156/169(old)/176/197/218. old code; don't need it anymore
172-192. Rewrote the search pattern: now goes forward, goes back to the
middle and searches backwards, and then quits in the middle.

227-238. this is the biggest addition to the code: a brand new algorithm
that makes the hexapod run so much better. This algorithm balances out
the robot fluidly as it walks. Here's how it works:
if the foot is not stepping back up (step 4 of its walk phase)
AND it is touching the ground
AND if it is higher or lower than the threshold range (3 units from the
middle position)
THEN it will increase by one unit every frame until it's in the middle
Possible fixes:
-Have it move exponentially towards the center so that it corrects
faster the further off balance it is

517-522. Reordered the legs so that it travels LR, RR, LM, RM, LF, RF;
keeps the hexapod more stable as it walks.
525. makes the legs move a little bit faster, but not so fast that the
sensors can't keep up.
542-547/550. same as above, but this is for the non-sensing gait.